+++ tests/test_core/test_readers/test_read_opus.py
@@ -11,6 +11,7 @@ import pytest
 import spectrochempy as scp
 from spectrochempy import NDDataset
 from spectrochempy import preferences as prefs
+from spectrochempy.utils.testing import assert_dataset_equal
 
 DATADIR = prefs.datadir
 OPUSDATA = DATADIR / "irdata" / "OPUS"
@@ -27,6 +28,17 @@ def test_read_opus():
     A = NDDataset.read_opus(OPUSDATA / "test.0000")
     assert A.shape == (1, 2567)
     assert A[0, 2303.8694].data == pytest.approx(2.72740, 0.00001)
+    assert A.units == "absorbance"
+
+    # background
+    B = NDDataset.read_opus(OPUSDATA / "background.0", type="RF")
+    assert B.shape == (1, 4096)
+    assert B.units is None
+
+    # Test if the background.0 file is correctly inferered as a reference spectrum
+    C = NDDataset.read_opus(OPUSDATA / "background.0")
+    assert C.shape == (1, 4096)
+    assert_dataset_equal(B, C)
 
     # read contents
     p = OPUSDATA / "test.0000"
@@ -35,4 +47,6 @@ def test_read_opus():
     assert F.name == p.name
     assert F.shape == (1, 2567)
 
-    assert NDDataset.read_opus(OPUSDATA / "background.0") is None
+
+if __name__ == "__main__":
+    pytest.main([__file__])